docs: reviewed directories and add pinia to dependencies
- delete src directory and move files to root
- render v-confirm on the same line as Layout
