+++ packages/docs/.vitepress/theme/plugins/i18n/locales/ja.ts
