+++ packages/docs/guide/component-api.md
