+++ packages/docs/ja/guide/component-api.md
