--- packages/docs/src/.vitepress/theme/components/LayoutProvider.vue
@@ -1,12 +1,0 @@
-<!-- https://github.com/vuejs/vitepress/blob/main/src/client/theme-default/Layout.vue -->
-<template>
-  <ClientOnly>
-    <VCLayout />
-  </ClientOnly>
-</template>
-
-<script setup lang="ts">
-import VCLayout from "./VCLayout.vue";
-</script>
-
-<style lang="sass" scoped></style>
